COMPLETELY CHANGED IT
Either I'm really smart and overthinking it or just a plain idiot, and I know I'm not smart. Spent hours and hours on this (2-3) and it still fails.
Submitted it in Python3 and got 0 points, Python2 got 9 points. I barely know the difference.
Inspired by http://mmhs.ca/ccc/2010/j32010.txt but obviously it's a different language.
Checked http://mmhs.ca/ccc/index.htm#2010 Problems
and my output looks like it matches the correct one, not sure what the difference is.
The question is,
where is my problem?